Adding and Removing Resources
Implemented some more interaction with the game screen and the blue buttons now have the interactability needed to add and remove resources. Note that this is still very heavily still a work in progress and has not been linked to the coins or matched up with the gray arrows yet.